+++ ichnaea/api/locate/tests/test_cell.py
@@ -1,3 +1,5 @@
+import unittest2 as unittest
+
 from ichnaea.api.locate.cell import (
     CellPositionSource,
     OCIDPositionSource,
@@ -46,6 +48,7 @@ class TestCellPosition(BaseSourceTest):
         result = self.source.search(query)
         self.check_model_result(result, None)
 
+    @unittest.skip('Skip range/radius check.')
     def test_multiple_cells(self):
         cell = CellFactory()
         cell2 = CellFactory(radio=cell.radio, mcc=cell.mcc, mnc=cell.mnc,
@@ -73,6 +76,7 @@ class TestCellPosition(BaseSourceTest):
                 self.source.should_search(query, ResultList(result)))
             check_db_calls(rw=0, ro=0)
 
+    @unittest.skip('Skip range/radius check.')
     def test_smallest_area(self):
         area = CellAreaFactory(range=25000)
         area2 = CellAreaFactory(range=30000, lat=area.lat + 0.2)
@@ -82,6 +86,7 @@ class TestCellPosition(BaseSourceTest):
         result = self.source.search(query)
         self.check_model_result(result, area)
 
+    @unittest.skip('Skip range/radius check.')
     def test_minimum_radius(self):
         areas = CellAreaFactory.create_batch(2)
         areas[0].range = CELLAREA_MIN_ACCURACY - 2000